Update Spectra_Moisture_Predictor.R
Changes to make code compliant with new directory format.

Changes to make compliant with new format of training dataset.  Was getting moisture values of 2-2.5, but after changing the columns subset, we got back to the 0.4-0.5 range.

Added a plot to look at the dispersion of predicted values, and the relative locations of samples chosen for validation tests.